Rename AppBasicCard with AppCard
