+++ components/ui/AppCard.vue
