+++ pages/events/index.vue
@@ -62,13 +62,13 @@
           />
         </div>
         <div v-else class="event-page__results">
-          <AppBasicCard
+          <AppCard
             class="event-page__no-events-card"
             :image="emptyCard.img"
             :title="emptyCard.title"
           >
             {{ emptyCard.description }}
-          </AppBasicCard>
+          </AppCard>
           <p class="event-page__no-events-msg">
             Nothing here yet -
             <AppLink
@@ -90,7 +90,7 @@ import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/logic/QiskitPage.vue'
 import EventCard from '~/components/events/EventCard.vue'
 import AppLink from '~/components/ui/AppLink.vue'
-import AppBasicCard from '~/components/ui/AppBasicCard.vue'
+import AppCard from '~/components/ui/AppCard.vue'
 import TheEventsHeader from '~/components/events/TheEventsHeader.vue'
 
 import {
@@ -104,7 +104,7 @@ import { EVENT_REQUEST_LINK } from '~/constants/appLinks'
   components: {
     EventCard,
     AppLink,
-    AppBasicCard,
+    AppCard,
     TheEventsHeader
   },
