+++ pages/learn/index.vue
@@ -21,7 +21,7 @@
         :compact="isShowingMoreResources"
         url="/learn/?learnLevel=Advanced&amp;timeScale=1%20minute"
       />
-      <AppBasicCard
+      <AppCard
         v-for="resource in filteredLearningResources"
         :key="resource.path"
         :image="resource.image"
@@ -30,7 +30,7 @@
         :cta-label="resource.ctaLabel"
       >
         <nuxt-content class="copy" :document="resource" />
-      </AppBasicCard>
+      </AppCard>
     </TheLearningResourceList>
   </div>
 </template>
@@ -41,7 +41,7 @@ import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/logic/QiskitPage.vue'
 import TheLearnHeader from '~/components/learn/TheLearnHeader.vue'
 import TheLearningResourceList from '~/components/learn/TheLearningResourceList.vue'
-import AppBasicCard from '~/components/ui/AppBasicCard.vue'
+import AppCard from '~/components/ui/AppCard.vue'
 import TheCarefulExplanationForBeginners from '~/components/learn/TheCarefulExplanationForBeginners.vue'
 import TheCarefulExplanationForExperts from '~/components/learn/TheCarefulExplanationForExperts.vue'
 import {
@@ -59,7 +59,7 @@ import {
     TheLearningResourceList,
     TheCarefulExplanationForExperts,
     TheCarefulExplanationForBeginners,
-    AppBasicCard
+    AppCard
   },
 
   head () {
